+++ src/pid.c
@@ -22,7 +22,7 @@ void pid_atualiza_referencia(float referencia_){
     referencia = (double) referencia_;
 }
 
-double pid_controle(double saida_medida){
+int pid_controle(double saida_medida){
     printf("Referencia: %f\n", referencia);
     printf("Temperatura Interna: %lf\n", saida_medida);
     double erro = referencia - saida_medida;
@@ -54,5 +54,5 @@ double pid_controle(double saida_medida){
     erro_anterior = erro;
     
     printf("Sinal de Controle: %lf\n", sinal_de_controle);
-    return sinal_de_controle;
+    return (int) sinal_de_controle;
 }
